[TASK] Migrate controller tests to functional tests (indexAction) (#1729)
It is considered best practice to cover controllers via functional
instead of unit tests.
We therefore migrate the existing unit tests to functional tests.

The migration is split into multiple batches for smaller changes.
This one covers tests related to `indexAction()`.

We can't add types to constants, as we still support PHP < 8.3.
We therefore completely disable this PHPStan issue for our project.

Relates: #1569