[TASK] Migrate controller tests to functional tests (newAction) (#1778)
It is considered best practice to cover controllers via functional
instead of unit tests.
We therefore migrate the existing unit tests to functional tests.

The migration is split into multiple batches for smaller changes.
This one covers tests related to `newAction()`.

We remove the `$tea` argument from `newAction()`, as the extension does
not provide a way in itself to call this action with this argument.

Relates: #1569