[TASK] Move "composer unused" configuration from root to Build folder
The project root should be as clean as possible.
Configuration files for tools should be within a dedicated folder like
`Build/`.
That's why we also move the configuration file for composer-unused.

This has the drawback that the tool can't be used with defaults, but
explicitly needs to be called with `--configuration` option.

Resolves: #1782