[TASK] Move `rector.php` configuration from project root to `Build/` (#1785)
The project root should be as clean as possible.
Configuration files for tools should be within a dedicated folder like
`Build/`.
That's why we also move the configuration file for rector.

This has the drawback that the tool can't be used with defaults, but
explicitly needs to be called with `--config` option.

Resolves: #1782