[TASK] Move eslint configuration from project root to `Build/`
The project root should be as clean as possible.
Configuration files for tools should be within a dedicated folder like
`Build/`.
That's why we also move the configuration file for eslint.

This has the drawback that the tool can't be used with defaults, but
explicitly needs to be called with `--config` option.

Resolves: #1788