[FEATURE] Add functional tests to update properties of teas
We already had a test covering the update of the title.

But the process also allows to update the description.
That's why we add another test covering this part of the process.
We adjust the existing fixture and test name to reflect that this is a
more specific test case.

Right now we don't offer to update other properties through the process.

Resolves: #1793