[TASK] Remove unused `.prettierrc.js` file from project root (#1795)
We want to clean up the project root.
We couldn't find a usage of the file. Oliver tested the plugin with his
IDE and noticed that it didn't seem to work, e.g. a rule was not
applied, everything else was applied even if the file didn't exist.

We therefore now remove the file.

Resolves: #1791