[TASK] Unblock argument passing in `Build/Scripts/runTests.sh` (#1689)
* [TASK] Unblock argument passing in `Build/Scripts/runTests.sh`

The first argument after processing defined options has been
stored into the variable `TEST_FILE`. This value was passed
to PHPUnit based commands as last argument, defining the test
case to execute (or a subset).

Removing this workaround and directly passing remaining options
to the command makes special arguments like `-e` obsolete and
unblocks more generic command passing.

This change removes this variable and passes all remaining
arguments after option processing directly to the PHPUnit
command and paves the way for future command improvements.

Note:   This makes the `-e` option basically obsolete, but
        is kept for now to minimize muscle memory issues.
        Using it will display a `deprecation` notice at the
        end of the script - so it gets recognized.

This is the Tea port of this change in the TYPO3 Core,
but without test chunking, the deprecated unit or functional
tests and without the acceptance tests (as we don't have
any of those):

https://review.typo3.org/c/Packages/TYPO3.CMS/+/82728

This is the first part of #1345.

* [BUGFIX] refactor back to using array as own variable

Related: #1689

---------

Co-authored-by: Eike Starkmann <eike.starkmann@posteo.de>