[TASK] Improve the Composer script names
As long as we still are using the Composer scripts, the names
should not be misleading.

We have had scripts with the `ci:` prefix for a long time.

However, these scripts are not only for CI, but generally for
checks that can (mostly) also be run locally.

So this prefix now has been changed to `check:` to be less
misleading.