[TASK] Thin out the CI matrix
Having CI jobs for the unit and functional tests with
PHP versions between the lowest and highest versions we support
does not add much data, but it incurs quite some energy cost
and increases the overall build time.

So remove these jobs.

We keep the jobs when the job with the next-highest PHP version
is only limited to the highest or lowest dependencies (as this
still provides as with valuable data).

Also keep the the PHP linting jobs at all versions as those are
quite fast and provide helpful feedback like "you are using a
PHP feature that was only introduced in PHP 8.2.".