[TASK] Use ErrorController for proper access denied (#1845)
We used a generic `\RuntimeException` in case of permission issues.
We now use the TYPO3 `ErrorController` API to build a proper
access denied response. This will take TYPO3 configuration into account
and return a proper HTTP status code.

The tests needed adjustments as we now can and should verify the actual
response, not a thrown exception anymore.

Resolves: #1844