[TASK] Drop the setup for GitLab CI (#1869)
Our GitLab CI setup has not been receiving the proper love and
maintenance that it deserved. One of the reasons is that we
already are using GitHub Actions for CI.

As one of the main reasons for this project to exist is to
provide current best-practice example that can be copied, our
GitLab integration did not live up to this expectation.

Also, we want to avoid having to maintain two parallel CI
setups.

So, at our team call on 2025-11-10, we decided to drop our
GitLab CI setup.

Fixes #1868
Closes #1852
Closes #1471
Closes #1141
Closes #1079
Closes #970
Closes #790
Closes #161